java21 examp;le added

+++ APPLICATION/JDK8/java8-customobj-sorting/src/main/java/com/espark/adarsh/bean/Product.java
@@ -10,6 +10,6 @@ public class Product {
 
     String categoryId;
     String productName;
-    Float productPrice;
+    Double productPrice;
 
 }
